Formalize the cross window message channel
This adds some specs and possibly over-engineers the back and forth
of the local editing context and the "remote" one within the frame.

This is arguably the basis for some of the concepts of our data-layer
work, but built by hand without the benefit of Q-connection for the
moment.

I'm fleshing it out this way and will clean up later.