Improve connection management when switching pages
Prior to this we relied too hazily on the currentPage as the trigger to
connect or disconnect. In reality, we only need to do this when the 
pageView's iframe actually unloads the live DOM of a page to which we
had a channel open to the agent inside the frame.

This only happens on unload, not when we change the currentPage.

The agent now posts messages acknowledging that the channel is open,
which eases some of the management locally as we know when we should
actually flush the queue of messages we tried to send when there was
no connection.

Additionally it also posts a message when the frame its in is unloading
its content to fix the aforementioned problem of knowing when to
disconnect.